Scan Function Online! (v0.3)
I Built the necessary GUI elements to accommodate the back-end of the scan function and made some edits to DirectoryList in order to join them together.

In addition to this I updated Main so that it launches the program and also unpackaged and re-factored my GUI classes to re-join the main project.

This Branch is now functional!